Create an interface for AnalyticsManager
This will help with testing later.

COAND-844
